Statistics: clean up unnecessary cast; analyzer clean

+++ lib/pages/statistics_page.dart
@@ -130,7 +130,7 @@ class _StatisticsPageState extends State<StatisticsPage> {
           final rts = await api.fetchSeasonEpisodeRuntimes(show.id, season.seasonNumber);
           int sum = 0;
           for (var i = 0; i < w; i++) {
-            final int rt = (i < rts.length ? (rts[i] as int) : 0);
+            final int rt = (i < rts.length ? rts[i] : 0);
             if (rt > 0) {
               sum += rt;
             } else {
